pages for courses were made

+++ src/main/java/com/example/javarush/Repository/CourseRepository.java
@@ -3,6 +3,8 @@ package com.example.javarush.Repository;
 import com.example.javarush.Entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
+
+
 
 }
