+++ src/main/java/com/example/javarush/Repository/ModulesRepository.java
@@ -3,7 +3,7 @@ package com.example.javarush.Repository;
 import com.example.javarush.Entity.Modules;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModulesRepository extends JpaRepository<Modules, Integer> {
+public interface ModulesRepository extends JpaRepository<Modules, Long> {
 
 
 }
